[velero]: be consistent: add namespace vars everywhere
This is "controversial" as some people like not specifying the namespace
on the helm definitions and instead rely on specifying it afterwards
doing a `helm template | kubectl -n foo create -f -`.

However, as some resources definitely need a namespace
(ClusterRoleBinding for example), thus the namespace *needs* to be specified
on both commands `helm template --namespace foo | kubectl -n foo` which
creates redundancy.

In order to be consistent, we specify the `namespace` param on every
resource. This was already the case in the
[charts/velero/templates/update-crds.yaml] file, so we just make it
consistent across every resource.

Signed-off-by: Frank Villaro-Dixon <frank.villaro@infomaniak.com>