Update example PrometheusRule
The provided example shall never be one that gives false negatives on the subject of whether a backup is failing. The consequences may be catastrophic if someone does not notice it is an absolutely horrible example that is giving a false sense of security.

It would be better to remove the example altogether than to give one that can keep claiming everything is in order when there is a chain of failed backups instead

Signed-off-by: Marcin Cieślak <marcin.k.cieslak@gmail.com>